This is my senior project that i did at DeVry University for my senior year.
It is a HVAC zone control project that allowed the user to control the
temp of each room individually. I made a Desktop, Web and an Android App
that allowed the user to have access to its thermostat. The languages
that were used were Java(Servlets, JDBC), C++ , HTML, CSS, JavaScript.